refactor : change package (club code) with code fomatting
