+++ api/src/test/java/com/hcs/club/ClubControllerTest.java
@@ -1,9 +1,9 @@
-package com.hcs.controller.club;
+package com.hcs.club;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hcs.domain.Club;
 import com.hcs.dto.ClubDto;
-import com.hcs.mapper.ClubMapper;
+import com.hcs.mapper.club.ClubMapper;
 import com.ulisesbocchio.jasyptspringboot.annotation.EnableEncryptableProperties;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
