+++ api/src/test/java/com/hcs/club/ClubMapperTest.java
@@ -1,7 +1,8 @@
-package com.hcs.mapper;
+package com.hcs.club;
 
 import com.hcs.domain.Club;
 import com.hcs.domain.User;
+import com.hcs.mapper.club.ClubMapper;
 import com.hcs.mapper.user.UserMapper;
 import com.ulisesbocchio.jasyptspringboot.annotation.EnableEncryptableProperties;
 import org.junit.jupiter.api.DisplayName;
